requirements.txt
Flask
flask-bcrypt
PyJWT
requests